introduce abstract KeyWrapProvidr
